bug fixed in decoding:  if syncword found in Xing VBR header toc, it would confuse things.
Also fixed overflow in vbrtag.c, and changed size of xing vbr
header to 128kbs.  This will make non-vbr aware players
give a more reasonable playing time estimate.

+++ lame.h
@@ -185,8 +185,8 @@ typedef struct  {
   BRHST brhist;					  /* histogramming data */
 
   int version;                    /* 0=MPEG2  1=MPEG1 */
-  long int frameNum;              /* frame counter */
-  long totalframes;               /* frames: 0..totalframes-1 (estimate)*/
+  int frameNum;                   /* frame counter */
+  int totalframes;                /* frames: 0..totalframes-1 (estimate)*/
   int encoder_delay;
   int framesize;
